Try to tune a bunch of stuff, correctly show error status. SignalStrength seems wonky.
I'm getting good runs in the sense that the usb device
remains responsive and I can talk to it using many
commands.  However, it's clear that the protocol as
implemented isn't quite correct.  The responses for signal
strength are kind of unexpected, but at least the errors
seem pretty consistent.