subcommands and a flaky clmm driver
Here we have "working" subcommands, exposing a very flaky
clmm driver.  The clmm driver doesn't run cleanly over
subsequent runs.  Moreover, when using the automatic port
scanning feature, the protocol flow is unable to complete
successfully, even though ostensibly the same flow
executes fine when invoked directly and when the port is
specified.

Eg, the second one is flaky in the sense that it sometimes
does not complete successfully.  Repeated attempts
generally work unless the device is wedged.  Given that it
seems to work most of the time in the second form, the
first one should be the same code.  It regularly dies when
validating either ACK or readStatus (reading the command
header) when using the auto port scanning feature.

:::
  # fails
  PYTHONPATH=src/ python -m insulaudit.main -v clmm   hello
  # seems to work when retrying:
  PYTHONPATH=src/ python -m insulaudit.main -v clmm --port /dev/ttyUSB0  hello
  PYTHONPATH=src/ python src/insulaudit/devices/clmm/proto.py /dev/ttyUSB0

Suggests some kind of timing issue, possibly having to do
with the use of select by pyserial??

In any case, this error is reliably generated by the first form. :-(
, line 370, in getNumBytesAvailable
    result = self.readStatus( )
  File "/home/bewest/Documents/git/insulaudit/src/insulaudit/devices/clmm/proto.py", line 397, in readStatus
    assert commStatus == 0, ("command status not 0: %s" % (commStatus))
AssertionError: command status not 0: 189