Stub out project details: decode historical data
Here are a bunch of files which represent historical data
read from a pump.  The serial number is in code.txt.  The
data comes from Andy Rozman, I believe it was generated
using a test pump not in production use, and is safe to
publish.

The goal is to write a bunch of scripts able to verify the
contents of the csv using the *.data files as input.